+++ cvat-core/src/server-proxy.js
@@ -554,12 +554,12 @@
 
             async function exportTask(id) {
                 const { backendAPI } = config;
-                const url = `${backendAPI}/tasks/${id}`;
+                const url = `${backendAPI}/tasks/${id}/backup`;
 
                 return new Promise((resolve, reject) => {
                     async function request() {
                         try {
-                            const response = await Axios.get(`${url}?action=export`, {
+                            const response = await Axios.get(url, {
                                 proxy: config.proxy,
                             });
                             if (response.status === 202) {
@@ -585,7 +585,7 @@
                 return new Promise((resolve, reject) => {
                     async function request() {
                         try {
-                            const response = await Axios.post(`${backendAPI}/tasks?action=import`, taskData, {
+                            const response = await Axios.post(`${backendAPI}/tasks/backup`, taskData, {
                                 proxy: config.proxy,
                             });
                             if (response.status === 202) {
@@ -595,6 +595,59 @@
                             } else {
                                 const importedTask = await getTasks(`?id=${response.data.id}`);
                                 resolve(importedTask[0]);
+                            }
+                        } catch (errorData) {
+                            reject(generateError(errorData));
+                        }
+                    }
+
+                    setTimeout(request);
+                });
+            }
+
+            async function backupProject(id) {
+                const { backendAPI } = config;
+                const url = `${backendAPI}/projects/${id}/backup`;
+
+                return new Promise((resolve, reject) => {
+                    async function request() {
+                        try {
+                            const response = await Axios.get(url, {
+                                proxy: config.proxy,
+                            });
+                            if (response.status === 202) {
+                                setTimeout(request, 3000);
+                            } else {
+                                resolve(`${url}?action=download`);
+                            }
+                        } catch (errorData) {
+                            reject(generateError(errorData));
+                        }
+                    }
+
+                    setTimeout(request);
+                });
+            }
+
+            async function restoreProject(file) {
+                const { backendAPI } = config;
+
+                let data = new FormData();
+                data.append('project_file', file);
+
+                return new Promise((resolve, reject) => {
+                    async function request() {
+                        try {
+                            const response = await Axios.post(`${backendAPI}/projects/backup`, data, {
+                                proxy: config.proxy,
+                            });
+                            if (response.status === 202) {
+                                data = new FormData();
+                                data.append('rq_id', response.data.rq_id);
+                                setTimeout(request, 3000);
+                            } else {
+                                const restoredProject = await getProjects(`?id=${response.data.id}`);
+                                resolve(restoredProject[0]);
                             }
                         } catch (errorData) {
                             reject(generateError(errorData));
@@ -1476,6 +1529,8 @@
                             create: createProject,
                             delete: deleteProject,
                             exportDataset: exportDataset('projects'),
+                            backupProject,
+                            restoreProject,
                             importDataset,
                         }),
                         writable: false,