+++ tests/cypress/integration/actions_tasks2/case_97_export_import_task.js
@@ -83,7 +83,7 @@ context('Export, import an annotation task.', { browser: '!firefox' }, () => {
             cy.get('.ant-dropdown')
                 .not('.ant-dropdown-hidden')
                 .within(() => {
-                    cy.contains('[role="menuitem"]', new RegExp('^Export task$')).click().trigger('mouseout');
+                    cy.contains('[role="menuitem"]', new RegExp('^Backup Task$')).click().trigger('mouseout');
                 });
             cy.getDownloadFileName().then((file) => {
                 taskBackupArchiveFullName = file;
@@ -93,7 +93,7 @@ context('Export, import an annotation task.', { browser: '!firefox' }, () => {
         });
 
         it('Import the task. Check id, labels, shape.', () => {
-            cy.intercept('POST', '/api/v1/tasks?action=import').as('importTask');
+            cy.intercept('POST', '/api/v1/tasks/backup').as('importTask');
             cy.get('.cvat-import-task').click().find('input[type=file]').attachFile(taskBackupArchiveFullName);
             cy.wait('@importTask', { timeout: 5000 }).its('response.statusCode').should('equal', 202);
             cy.wait('@importTask').its('response.statusCode').should('equal', 201);